Struct rsp2_structure_io::assemble::RawAssemble

pub struct RawAssemble {
    pub normal_axis: usize,
    pub lattice: Lattice,
    pub fracs_in_plane: Vec<Vec<V3>>,
    pub carts_along_normal: Vec<Vec<f64>>,
    pub initial_vacuum_sep: f64,
    pub initial_layer_seps: Vec<f64>,
    pub initial_scale: Option<[f64; 3]>,
    pub part: Option<Part<usize>>,
    pub check_intralayer_distance: Option<f64>,
}

Fields
normal_axis: usize

The axis normal to the layers, as the index of a lattice vector.

lattice: Lattice

A lattice in which the lattice vector for normal_axis is orthogonal to the other two. The length of the normal_axis vector is ignored.

fracs_in_plane: Vec<Vec<V3>>

Fractional coordinates for each layer. Must all be zero in the normal axis.

carts_along_normal: Vec<Vec<f64>>

Cartesian coordinates along the normal within each layer. This allows layers to have a “thickness” to them.

Only the variation in value within each layer matters. It is important that, within each layer, these coordinates form a contiguous image.

initial_vacuum_sep: f64

Initial separation across periodic boundary.

initial_layer_seps: Vec<f64>

Initial separations between layers.

initial_scale: Option<[f64; 3]>

Initial scale factors. The value on normal_axis is ignored.

part: Option<Part<usize>>

The partition that decided the order of the atoms in fracs_in_plane.

If provided, it is used to make sure that the order of the atoms in the assembled structure matches the original order of the structure that was partitioned.

check_intralayer_distance: Option<f64>

Adds a sanity check that each pair of successive atoms along a layer lie within this distance from each other (with a little extra bit of fuzz). Construction of an Assemble will fail if this is untrue.

This is to help catch bugs related to accidentally wrapping carts_along_normal across a periodic boundary.